[#240] 사용자 가이드 하이라이팅 컴포넌트가 모달에서도 하이라이팅 (#244)
* 🐛 fix: 사용자 가이드 도중에 닫는 경우 상태 초기화

* 🐛 fix: 툴박스 관련 로직 코치마크로 이동

* 🙀 chore: TODO 삭제

* 💄 design: 코드 내보내기 버튼 배경색 추가

* 💄 design: 도움말 버튼 hover시 색상 좀 더 진하게 적용
