+++ apps/client/src/widgets/workspace/CoachMark/CoachMark.tsx
@@ -1,16 +1,13 @@
-import { useEffect } from 'react';
 import { useCoachMarkStore } from '@/shared/store/useCoachMarkStore';
 import { CircleButton } from '@/shared/ui';
+import { coachMarkContent } from '@/shared/utils';
+import { useEffect } from 'react';
 import * as Blockly from 'blockly/core';
 import TabbedToolbox from '@/core/tabbedToolbox';
-import { coachMarkContent } from '@/shared/utils';
 
 export const CoachMark = () => {
   const { currentStep, setCurrentStep, closeCoachMark } = useCoachMarkStore();
   const stepsLength = coachMarkContent.length;
-  const toolboxDiv = document.querySelector('.blocklyToolboxDiv');
-  // const blockCanvas = document.querySelector('.blocklyBlockCanvas');
-  // TODO: 사용자가이드 - 블록 하이라이팅
 
   const nextStep = () => {
     if (currentStep < stepsLength - 1) {
@@ -36,7 +33,6 @@ export const CoachMark = () => {
     }
 
     const toolbox = workspace.getToolbox() as TabbedToolbox;
-
     if (!toolbox) return;
 
     switch (currentStep) {
@@ -49,17 +45,7 @@ export const CoachMark = () => {
       case 2:
         toolbox.clickTab('html');
     }
-
-    if (toolboxDiv) {
-      if (currentStep <= 1) {
-        toolboxDiv.classList.add('coachMarkHighlight');
-        // blockCanvas.classList.add('coachMarkHighlight');
-      } else {
-        toolboxDiv.classList.remove('coachMarkHighlight');
-        // blockCanvas.classList.remove('coachMarkHighlight');
-      }
-    }
-  }, [currentStep, toolboxDiv]);
+  }, [currentStep]);
 
   return (
     <div className="fixed inset-0 z-[100] flex items-center justify-center bg-black bg-opacity-70">
@@ -83,7 +69,6 @@ export const CoachMark = () => {
               그만 보기
             </CircleButton>
 
-            {/* TODO: 이전 버튼 - 기능 구현 후 삭제 */}
             <CircleButton
               className="text-bold-sm h-8 w-16"
               onClick={prevStep}
